Build dataset using COPY instead of multi-row inserts
The fastest way to bulk insert data in Postgres is by using COPY. This
changes the dataset building to use that. I tried building a dataset for
1000 warehouses using 100 vusers. Without copy this took 100 minutes,
with copy it only took 42 minutes. So it reduced the time it takes to
build the dataset by ~58%.

Docs on the usage of COPY in the tcl Postgres library can be found here:
http://pgtclng.sourceforge.net/pgtcldocs/pgtcl-example-copy.html