Add optional Citus support (#302)
* Standardize indentation of tcl files

The tcl files in this repo did not seem to adhere to any common
indentation style. This made it quite hard for me to understand what the
code was doing. To resolve this I used the reformat.tcl script that is
provided on the tcl wiki:
https://wiki.tcl-lang.org/page/Reformatting+Tcl+code+indentation

I replaced the original reformat proc that was provided in the first
message, with the reformat2 proc that was contributed by aplsimple in
the last message on the thread. I did make one final change to this
reformat2 script, such that does not exclude comments from indentation.

I think this makes the code much more readable.

I understand that this is a big PR, but it only changes whitespace. If
you look at the diff by excluding whitespace, there's only one change.
And this change is a change from file encoding from latin1 to utf-8.

* Add optional Citus support

This adds the option to configure the TPROC-C build step to build the
schema with support for Citus. This distributes the tables and all the
stored procedures.

* fix .tpc.f1.e11a unknown command in driver options (#4)

Co-authored-by: Steve Shaw <38044085+sm-shaw@users.noreply.github.com>