Repair ora_compatible OSTAT procedure
An extra EXCEPTION clause was inserted at some point.  This syntax error
caused the creation of the database to also miss the creation of the
SLEV stored procedure when using EPAS in Oracle compatibility mode with
stored procedures.

Remove the extra condition "WHEN NO_DATA_FOUND".

It appears to have be broken since before the git repository was
created.