Container file simplication + 4.8 version bump
To reduce the duplication and minimize the image size, base/Dockerfile
was created to capture the common elements of the base image.

The database specific files derive from this base image.

The top level Dockerfile uses the mssql, the more complicated image
as a base, and using multistage images takes the important aspects out
of oracle's image, and does the package installs for maria, mysql and
postgres.

A number of changes in the base image where made:
* Apt-sources weren't used so leave them disabled.
* apt doesn't provide a stable command line interfaces, so use apt-get
* to reduce image size,  /var/lib/apt/lists/* is removed
* DEBIAN_FRONTEND=noninteractive is an environment variable rather than
  a build arg.