Alter MSSQLS TPCH Schema to use Decimal not Money
Altered the data type of the MSSQL TPCH load to store monetary
values in a decimal (12, 2) format, as opposed to the money datatype.

The money datatype has a precision that is too high to use the Aggregate
pushdown mechanic offered in SQL Server. Altered the money datatypes to
a decimal type to improve the performance of queries run that using
aggregation.