MSSQLS TPCH Add Flag to Partition Orders and Lineitems
Added a flag that allows the Order and Lineitem tables to be
partitioned. Each partition spans 1 week. The parition flag is set to
false by default and must be enabled.

Column Store and Row store are allowed to be partitioned.

When paritioning is enabled. Certain primary and foreign key definitions
defined in HammerDB are no longer valid. HammerDB will still create
those keys/constraints when the Orders and Lineitem  tables are not
partitioned.`

For columnstore indices, the is an initial degredation of performance,
however if the user calls Alter Index ... Reoganize or leaves the server
running after the initial load for approximately 48 hours, the
performance of power runs improves by 2x when compared to column store
indices with no partitioning.